refactor: move Claude adapter files into builtin plugin, empty AdapterRegistry constructor
