+++ packages/core/src/plugins/builtin/claude/frontmatter.ts
