+++ packages/core/src/plugins/builtin/claude/skills.ts
